requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
cssselect==1.2.0
exceptiongroup==1.1.0
fake-useragent==1.1.1
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
lxml==4.9.2
numpy==1.24.1
outcome==1.2.0
pandas==1.5.2
parse==1.19.0
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
requests-html==0.10.0
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
w3lib==2.1.1
websockets==10.4
wsproto==1.2.0
zipp==3.11.0
